requirements.txt
python-dotenv
streamlit
tavily-python
requests
google-genai